feat(infra): replace use of BigQuery by Firestore to store run configurations

+++ scripts/get-spirv-tools.sh
@@ -2,15 +2,16 @@
 cd bin
 [ $? -eq 0 ] && wget -nd -m https://storage.googleapis.com/spirv-tools/badges/build_link_$1_clang_release.html
 [ $? -eq 0 ] && tar -xvf install.tgz
+[ $? -eq 0 ] && rm spirv-*
 [ $? -eq 0 ] && cp install/bin/spirv-* .
-rm *.html
-rm install.tgz
-rm -rf install
-git clone https://github.com/KhronosGroup/SPIRV-Cross.git
+[ $? -eq 0 ] && rm *.html
+[ $? -eq 0 ] && rm install.tgz
+[ $? -eq 0 ] && rm -rf install
+[ $? -eq 0 ] && git clone https://github.com/KhronosGroup/SPIRV-Cross.git
 [ $? -eq 0 ] && cd SPIRV-Cross
 [ $? -eq 0 ] && make
 [ $? -eq 0 ] && cp spirv-cross ../spirv-cross-bin
-cd ..
-rm -rf SPIRV-Cross
-mv spirv-cross-bin spirv-cross
+[ $? -eq 0 ] && cd ..
+[ $? -eq 0 ] && rm -rf SPIRV-Cross
+[ $? -eq 0 ] && mv spirv-cross-bin spirv-cross
 cd ..
